+++ backbones/interpretableresnet.py
@@ -2,68 +2,7 @@ from torch import nn
 import torch
 import pdb
 
-class ChannelAttention(nn.Module):
-    def __init__(self, in_planes, ratio=16):
-        super(ChannelAttention, self).__init__()
-        self.avg_pool = nn.AdaptiveAvgPool2d(1)
-        self.max_pool = nn.AdaptiveMaxPool2d(1)
-
-        self.fc1   = nn.Conv2d(in_planes, in_planes // 16, 1, bias=False)
-        self.relu1 = nn.ReLU()
-        self.fc2   = nn.Conv2d(in_planes // 16, in_planes, 1, bias=False)
-
-        self.sigmoid = nn.Sigmoid()
-
-    def forward(self, x):
-        avg_out = self.fc2(self.relu1(self.fc1(self.avg_pool(x))))
-        max_out = self.fc2(self.relu1(self.fc1(self.max_pool(x))))
-        out = avg_out + max_out
-        return self.sigmoid(out)
-
-class SpatialAttention(nn.Module):
-    def __init__(self, kernel_size=7):
-        super(SpatialAttention, self).__init__()
-
-        assert kernel_size in (3, 7), 'kernel size must be 3 or 7'
-        padding = 3 if kernel_size == 7 else 1
-
-        self.conv1 = nn.Conv2d(2, 1, kernel_size, padding=padding, bias=False)
-        self.sigmoid = nn.Sigmoid()
-
-    def forward(self, x):
-        avg_out = torch.mean(x, dim=1, keepdim=True)
-        max_out, _ = torch.max(x, dim=1, keepdim=True)
-        x = torch.cat([avg_out, max_out], dim=1)
-        x = self.conv1(x)
-        return self.sigmoid(x)
-
-class DenseAttention(nn.Module):    # Like spatial, but for all channels
-    def __init__(self, planes):
-        super(DenseAttention, self).__init__()
-        self.dense_conv1=nn.Conv2d(planes,planes,kernel_size=3,padding=1,bias=False)
-        self.dense_relu1=nn.LeakyReLU()
-        self.dense_conv2=nn.Conv2d(planes,planes,kernel_size=3,padding=1,bias=False)
-        self.dense_sigmoid = nn.Sigmoid()
-    def forward(self,x):
-        x = self.dense_conv1(x)
-        x = self.dense_relu1(x)
-        x = self.dense_conv2(x)
-        x = self.dense_sigmoid(x)
-        return x
-
-class InputAttention(nn.Module):
-    def __init__(self, planes):
-        super(InputAttention, self).__init__()
-        self.ia_conv1=nn.Conv2d(planes,planes,kernel_size=3,padding=1,bias=False)
-        self.ia_relu1=nn.LeakyReLU()
-        self.ia_conv2=nn.Conv2d(planes,planes,kernel_size=3,padding=1,bias=False)
-        self.ia_sigmoid = nn.Sigmoid()
-    def forward(self,x):
-        x = self.ia_conv1(x)
-        x = self.ia_relu1(x)
-        x = self.ia_conv2(x)
-        x = self.ia_sigmoid(x)
-        return x
+from utils.blocks import ChannelAttention, DenseAttention, InputAttention, SpatialAttention
 
 class BasicBlock(nn.Module):
     expansion = 1
